# Api-Backend-LocSystem/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocSystem API",
        "description": "Documentação da API do sistema LocSystem",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/advisory-users": {
            "get": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Lista os usuários de assessoria com suporte a filtros administrativos.",
                "operationId": "1c84b99ea44eaddb4c2713a77e202f4f",
                "responses": {
                    "200": {
                        "description": "Usuários de assessoria retornados com sucesso no campo advisoryUsers."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem consultar usuários de assessoria."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao listar os usuários de assessoria."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Cadastra um novo usuário de assessoria e vincula as assessorias informadas.",
                "operationId": "0d94d2bb567c3a56379fab18db383a57",
                "responses": {
                    "201": {
                        "description": "Usuário de assessoria cadastrado com sucesso e retornado no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem cadastrar usuários de assessoria."
                    },
                    "409": {
                        "description": "O e-mail informado já está em uso por outro usuário."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados ou assessoria informada inválida."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao cadastrar o usuário de assessoria."
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 20
                                    },
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 32
                                    },
                                    "legalAdvisoryIds": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "properties": []
                                        },
                                        "minItems": 1
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_phone",
                                    "v_password",
                                    "legalAdvisoryIds"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/advisory-users/{id}": {
            "get": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Retorna os detalhes de um usuário de assessoria específico.",
                "operationId": "60e5864fb2e2ce01cd59867875f8b807",
                "responses": {
                    "200": {
                        "description": "Usuário de assessoria retornado com sucesso no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao consultar o usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Atualiza os dados e os vínculos de assessorias de um usuário de assessoria.",
                "operationId": "4340cecbd4e02ce8e961311e3251c268",
                "responses": {
                    "200": {
                        "description": "Usuário de assessoria atualizado com sucesso e retornado no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "409": {
                        "description": "O e-mail informado já está em uso por outro usuário."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados ou assessoria informada inválida."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 20
                                    },
                                    "legalAdvisoryIds": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "properties": []
                                        },
                                        "minItems": 1
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_phone",
                                    "legalAdvisoryIds"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Remove um usuário de assessoria quando não houver veículos vinculados aos seus acessos.",
                "operationId": "7f130bf02b1179f3d5ee3e34d8457c9c",
                "responses": {
                    "200": {
                        "description": "Usuário de assessoria removido com sucesso e confirmação retornada no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "409": {
                        "description": "O usuário não pode ser removido porque possui veículos vinculados às assessorias acessíveis."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao remover o usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/advisory-users/{id}/approval": {
            "patch": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Aprova um usuário de assessoria para acesso ao sistema.",
                "operationId": "e9cb828146754f5f8be51829690e7d8c",
                "responses": {
                    "200": {
                        "description": "Usuário de assessoria aprovado com sucesso e retornado no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao aprovar o usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/advisory-users/{id}/status": {
            "patch": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Ativa ou bloqueia um usuário de assessoria.",
                "operationId": "dacfa6fce2bb0c4cc47cab42a3b8b5a7",
                "responses": {
                    "200": {
                        "description": "Status do usuário atualizado com sucesso e dados retornados no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "422": {
                        "description": "Erro de validação no payload de alteração de status."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o status do usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean"
                                    }
                                },
                                "required": [
                                    "isActive"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/advisory-users/{id}/reset-2fa": {
            "post": {
                "tags": [
                    "Usuários de Assessoria"
                ],
                "summary": "Reseta a configuração de autenticação em dois fatores de um usuário de assessoria.",
                "operationId": "553392599f7f2284fef83825db429984",
                "responses": {
                    "200": {
                        "description": "2FA resetado com sucesso e confirmação retornada no campo advisoryUser."
                    },
                    "401": {
                        "description": "Acesso negado ou usuário de assessoria não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao resetar o 2FA do usuário de assessoria."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}